[svn] Port over ratbox 2.2 r23507, r23624, r23626 (jilles/anfl): Change TS6 JOIN processing - don't send out simple modes in TS6 JOIN and TS5 SJOIN when   a local user joins an existing channel - don't send out simple modes in TS6 JOIN and TS5 SJOIN when   propagating a TS6 JOIN - don't interpret simple modes in an incoming TS6 JOIN
This is to avoid desyncs when certain mode changes (e.g. -im)
cross with joins. A downside is that simple modes will be
more desynched when a JOIN creates a channel or lowers TS,
but that's less important.

Update the TS6 specification to include this, and clarify
that TMODE can come from a server and that MODE must be
translated into TMODE from other servers too.